Moved more files around.
