+++ spec/spec_helper.rb
@@ -8,7 +8,7 @@ ROOT = Pathname.new(File.expand_path('../../', __FILE__))
 $LOAD_PATH.unshift((ROOT + 'lib').to_s)
 $LOAD_PATH.unshift((ROOT + 'spec').to_s)
 
-require 'playgroundbook_lint'
+require 'playgroundbook'
 require 'playgroundbook_lint/abstract_linter'
 require 'playgroundbook_lint/chapter_linter'
 require 'playgroundbook_lint/chapter_manifest_linter'
